+++ lib/ts/components/superTokensRoute.tsx
@@ -18,8 +18,8 @@
  */
 
 import * as React from "react";
-import NormalisedURLPath from "../normalisedURLPath";
 import SuperTokens from "../superTokens";
+import { RoutingComponent } from "./routingComponent";
 
 /*
  * Component.
@@ -37,27 +37,8 @@ export function getSuperTokensRoutesForReactRouterDom(supertokensInstance: Super
         path = path === "" ? "/" : path;
         return (
             <Route exact key={`st-${path}`} path={path}>
-                <SuperTokensRouteWithRecipeId supertokensInstance={supertokensInstance} path={path} />
+                <RoutingComponent supertokensInstance={supertokensInstance} path={path} />
             </Route>
         );
     });
-}
-
-function SuperTokensRouteWithRecipeId({
-    supertokensInstance,
-    path,
-}: {
-    supertokensInstance: SuperTokens;
-    path: string;
-}): JSX.Element | null {
-    const history = supertokensInstance.getReactRouterDom()?.useHistoryCustom();
-    const normalisedPath = new NormalisedURLPath(path);
-
-    const featureComponentWithRecipeId = supertokensInstance.getMatchingComponentForRouteAndRecipeId(normalisedPath);
-
-    if (featureComponentWithRecipeId === undefined) {
-        return null;
-    }
-
-    return <featureComponentWithRecipeId.component history={history} />;
 }