+++ test/end-to-end/thirdparty.test.js
@@ -113,7 +113,6 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS SESSION OVERRIDE ADD_FETCH_INTERCEPTORS_AND_RETURN_MODIFIED_FETCH",
                 "ST_LOGS SESSION OVERRIDE ADD_AXIOS_INTERCEPTORS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY OVERRIDE SIGN_IN_AND_UP",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY PRE_API_HOOKS THIRD_PARTY_SIGN_IN_UP",
@@ -123,8 +122,6 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS THIRD_PARTY ON_HANDLE_EVENT SUCCESS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY GET_REDIRECTION_URL SUCCESS",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS SESSION OVERRIDE GET_JWT_PAYLOAD_SECURELY",
                 "ST_LOGS SESSION OVERRIDE GET_USER_ID",
             ]);
@@ -213,12 +210,9 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS SESSION OVERRIDE ADD_FETCH_INTERCEPTORS_AND_RETURN_MODIFIED_FETCH",
                 "ST_LOGS SESSION OVERRIDE ADD_AXIOS_INTERCEPTORS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY OVERRIDE SIGN_IN_AND_UP",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY GET_REDIRECTION_URL SIGN_IN_AND_UP",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
             ]);
             const pathname = await page.evaluate(() => window.location.pathname);
             const search = await page.evaluate(() => window.location.search);
@@ -245,12 +239,9 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS SESSION OVERRIDE ADD_FETCH_INTERCEPTORS_AND_RETURN_MODIFIED_FETCH",
                 "ST_LOGS SESSION OVERRIDE ADD_AXIOS_INTERCEPTORS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY OVERRIDE SIGN_IN_AND_UP",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY GET_REDIRECTION_URL SIGN_IN_AND_UP",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
             ]);
             const pathname = await page.evaluate(() => window.location.pathname);
             const search = await page.evaluate(() => window.location.search);
@@ -277,12 +268,9 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS SESSION OVERRIDE ADD_FETCH_INTERCEPTORS_AND_RETURN_MODIFIED_FETCH",
                 "ST_LOGS SESSION OVERRIDE ADD_AXIOS_INTERCEPTORS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY OVERRIDE SIGN_IN_AND_UP",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY GET_REDIRECTION_URL SIGN_IN_AND_UP",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
             ]);
             const pathname = await page.evaluate(() => window.location.pathname);
             const search = await page.evaluate(() => window.location.search);
@@ -333,12 +321,9 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS SESSION OVERRIDE ADD_FETCH_INTERCEPTORS_AND_RETURN_MODIFIED_FETCH",
                 "ST_LOGS SESSION OVERRIDE ADD_AXIOS_INTERCEPTORS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY OVERRIDE SIGN_IN_AND_UP",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY GET_REDIRECTION_URL SIGN_IN_AND_UP",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
             ]);
             const pathname = await page.evaluate(() => window.location.pathname);
             const search = await page.evaluate(() => window.location.search);
@@ -365,12 +350,9 @@ describe("SuperTokens Third Party", function () {
                 "ST_LOGS SESSION OVERRIDE ADD_FETCH_INTERCEPTORS_AND_RETURN_MODIFIED_FETCH",
                 "ST_LOGS SESSION OVERRIDE ADD_AXIOS_INTERCEPTORS",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY OVERRIDE SIGN_IN_AND_UP",
                 "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
                 "ST_LOGS THIRD_PARTY GET_REDIRECTION_URL SIGN_IN_AND_UP",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
-                "ST_LOGS THIRD_PARTY OVERRIDE GET_OAUTH_STATE",
             ]);
             const pathname = await page.evaluate(() => window.location.pathname);
             const search = await page.evaluate(() => window.location.search);